+++ app/src/main/java/com/barryburgle/gameapp/ui/navigation/BottomNavigationItem.kt
@@ -1,11 +1,12 @@
 package com.barryburgle.gameapp.ui.navigation
 
-import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
 
 data class BottomNavigationItem(
     val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
+    val icon: Painter,
+    val modifier: Modifier,
     val hasNews: Boolean,
     val badgeCount: Int? = null,
     val destinationScreen: String,